requirements.txt
arrow==1.2.1
asgiref==3.5.0
blessed==1.19.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-filter==21.1
django-nine==0.2.5
django-picklefield==3.0.1
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djongo==1.3.6
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
packaging==21.3
PyJWT==2.6.0
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
redis==4.5.3
requests==2.26.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
dnspython==2.2.1
openai==0.27.8
pandas==1.5.2
openpyxl==3.0.10
beautifulsoup4==4.12.2
python-dotenv==0.21.0
gunicorn==20.1.0
celery==5.3.6
django-celery-beat==2.5.0
firebase-admin==6.1.0